[CHORE] Refactor models to have a BaseModel they inherit from
