+++ app/models/base_models.py
@@ -1,0 +1,13 @@
+import os
+import psycopg2
+from instance.config import app_config
+
+CURRENT_ENVIRONMENT = os.environ['ENV']
+CONN_STRING = app_config[CURRENT_ENVIRONMENT].CONNECTION_STRING
+
+class BaseModel(object):
+    '''answers class model'''
+    def __init__(self):
+        '''open database connections'''
+        self.conn = psycopg2.connect(CONN_STRING)
+        self.cursor = self.conn.cursor()
