+++ app/models/revoked_token_model.py
@@ -1,18 +1,9 @@
 '''app/models/revoked_token_model.py'''
-import os
 import psycopg2
-from instance.config import app_config
+from app.models.base_models import BaseModel
 
-CURRENT_ENVIRONMENT = os.environ['ENV']
-CONN_STRING = app_config[CURRENT_ENVIRONMENT].CONNECTION_STRING
-
-class RevokedTokens(object):
+class RevokedTokens(BaseModel):
     '''class representing revoked tokens model'''
-    def __init__(self):
-        '''set up class variables'''
-        self.conn = psycopg2.connect(CONN_STRING)
-        self.cursor = self.conn.cursor()
-
     def add_token_to_blacklist(self, json_token_identifier):
         '''add token to blacklist'''
         sql = "INSERT INTO revoked_tokens (json_token_identifier) VALUES (%s);"
